Rename profile and client classes for API
